Use a more coherent layout for Injection modules
An object in a module should never depends on objects that are injected
by the same module. This couples objects and makes using injection
on unit tests hard.
Due to this a new structure for the modules was created. Now, each
module contains the recipes for a "kind" of object. This way, the
Databases module will only create Databases. And one Database should
never depend on another. Same with the other modules, the Repositories
module will only inject Repositories. And one Repository should never
depend on another (as they are in the same hierarchical level). A
repository must always depend on data sources (like Room Daos or
Retrofit Services).