Preparations to migrate to the Kotlin DSL
Groovy allows to do ambiguous things we all end up doing, like skipping
the = symbol for asignments and the parenthesis for method invokations.
Let's get rid of all these bad practices that are not allowed in the
Kotlin DSL, so if and when we migrate to it we don't have to do all this
refactoring on the gradle scripts.