Refactor to adopt a pure clean architecture approach
With this, the application has a usecase centric approach.
The presentation layer depends on the domain layer, which
represents the bussiness.
The data layer uses the domain layer as a contract to determine
what it should provide. Again, the domain layer represents the
bussiness.
With this, everything relies on bussiness rules, which are totally
agnostic to the rest of the implementation.